+++ zsh/plugins/syntax-highlighting/highlighters/main/test-data/meta-no-eval2.zsh
@@ -1,6 +1,6 @@
 #!/usr/bin/env zsh
 # -------------------------------------------------------------------------------------------------
-# Copyright (c) 2016 zsh-syntax-highlighting contributors
+# Copyright (c) 2020 zsh-syntax-highlighting contributors
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without modification, are permitted
@@ -28,12 +28,13 @@
 # vim: ft=zsh sw=2 ts=2 et
 # -------------------------------------------------------------------------------------------------
 
-# see alias-comment1.zsh
-setopt interactivecomments
-BUFFER=$'# foo\ntrue'
+# We aren't testing how this is highlighted; we're testing that it's not
+# evaluated.  If it gets evaluated, the test suite will die.
+BUFFER=$': /(e*exit 42*)'
 
 expected_region_highlight=(
-  '1 5 comment' # # foo
-  '6 6 commandseparator' # \n
-  '7 10 builtin' # true
+  '1 1 builtin' # :
+  '3 15 default' # /(e*exit 42*)
+  '6 6 globbing' # *
+  '14 14 globbing' # *
 )