+++ zsh/plugins/syntax-highlighting/highlighters/main/test-data/redirection-special-cases.zsh
@@ -1,6 +1,6 @@
 #!/usr/bin/env zsh
 # -------------------------------------------------------------------------------------------------
-# Copyright (c) 2016 zsh-syntax-highlighting contributors
+# Copyright (c) 2019 zsh-syntax-highlighting contributors
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without modification, are permitted
@@ -28,12 +28,17 @@
 # vim: ft=zsh sw=2 ts=2 et
 # -------------------------------------------------------------------------------------------------
 
-# see alias-comment1.zsh
-setopt interactivecomments
-BUFFER=$'# foo\ntrue'
+# See xpandredir() in the zsh source.
+
+BUFFER=$'cat <&p; exec {myfd}>&-'
 
 expected_region_highlight=(
-  '1 5 comment' # # foo
-  '6 6 commandseparator' # \n
-  '7 10 builtin' # true
+  '1 3 command' # cat
+  '5 6 redirection' # <&
+  '7 7 redirection' # p
+  '8 8 commandseparator' # ;
+  '10 13 precommand' # exec
+  '15 20 named-fd' # {myfd}
+  '21 22 redirection' # >&
+  '23 23 redirection' # -
 )